# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BudgetPayload:
    properties:
      currency:
        type: string
      period_end:
        type: string
      period_start:
        type: string
      scope_id:
        type: string
      scope_type:
        type: string
      total_limit:
        type: number
    type: object
  dto.DepartmentPayload:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
  dto.ExpenseCategoryPayload:
    properties:
      code:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
  dto.ProjectPayload:
    properties:
      code:
        type: string
      department_id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
  dto.UserPayload:
    properties:
      department_id:
        type: string
      email:
        type: string
      full_name:
        type: string
      is_active:
        type: boolean
      manager_id:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: REST API for the Business Trip and Expense Portal
  title: Travel Portal API
  version: "1.0"
paths:
  /admin/budgets:
    get:
      parameters:
      - description: department|project
        in: query
        name: scope_type
        type: string
      - description: Scope ID
        in: query
        name: scope_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List budgets
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Budget payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.BudgetPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create budget
      tags:
      - Admin
  /admin/categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Category payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ExpenseCategoryPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - Admin
  /admin/categories/{id}:
    delete:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - Admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ExpenseCategoryPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - Admin
  /admin/departments:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List departments
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Department payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.DepartmentPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create department
      tags:
      - Admin
  /admin/departments/{id}:
    delete:
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete department
      tags:
      - Admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: string
      - description: Department payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.DepartmentPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update department
      tags:
      - Admin
  /admin/projects:
    get:
      parameters:
      - description: Filter by department
        in: query
        name: department_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Project payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ProjectPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - Admin
  /admin/projects/{id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - Admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ProjectPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - Admin
  /admin/users:
    get:
      parameters:
      - description: Filter by role
        in: query
        name: role
        type: string
      - description: Filter by department
        in: query
        name: department_id
        type: string
      - description: Include inactive users
        in: query
        name: include_inactive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: User payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Admin
  /admin/users/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"